+++ R/utils.R
@@ -54,10 +54,6 @@ get_file_mtimes <- function(files) {
   )
 }
 
-"%||%" <- function(x, y) {
-  if (is.null(x)) y else x
-}
-
 # Checks whether a package is installed
 is_installed <- function(package) {
   nzchar(system.file(package = package))
@@ -127,6 +123,22 @@ collapse0 <- function(..., collapse = "\n") {
 #' @importFrom  rlang have_name
 has_any_name <- function(x) {
   any(have_name(x))
+}
+
+is_string <- function(x) {
+  is.character(x) && length(x) == 1
+}
+
+trim_ws <- function(x) {
+  sub("^\\s*", "", sub("\\s*$", "", x))
+}
+
+is_available <- function(package, version = NULL) {
+  installed <- nzchar(system.file(package = package))
+  if (is.null(version)) {
+    return(installed)
+  }
+  installed && isTRUE(utils::packageVersion(package) >= version)
 }
 
 has_any_name_recursive <- function(x) {