+++ tests/testthat/test-extend.R
@@ -1,5 +1,3 @@
-context("extend")
-
 # Disable sass cache
 local_disable_cache()