+++ tests/testthat/test-include-paths.R
@@ -1,5 +1,3 @@
-context("inclue paths")
-
 # Disable sass cache
 local_disable_cache()