+++ tests/testthat/test-variables.R
@@ -1,5 +1,3 @@
-context("variables")
-
 # Disable sass cache
 local_disable_cache()